requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==4.2.9
django-allauth==0.60.1
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
mysqlclient==2.2.3
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
